requirements.txt
bcrypt
beautifulsoup4
cryptography>=3.4.7
flask
google-auth
google-auth-oauthlib
numpy
pandas
PyJWT
requests
supabase
transformers
psycopg2-binary
Flask-SQLAlchemy
flask-login